Remove Circular Dependency
