+++ index.js
@@ -1,7 +1,7 @@
 import domSelectors from './modules/domSelectors.js';
 import Store from './modules/LocalStorage.js';
 import Book from './modules/Book.js';
-import UI from './modules/UI.js';
+// import UI from './modules/UI.js';
 
 setInterval(() => {
   document.getElementById('date').textContent = new Date();
@@ -28,6 +28,38 @@ domSelectors.GoToContact.addEventListener('click', () => {
   domSelectors.addBookView.classList.remove('show');
   domSelectors.bookListView.classList.remove('show');
 });
+
+// ------  UI CLASS ----- //
+
+class UI {
+  // Display the list of Books
+  static displayBooks() {
+    const books = Store.getBooks();
+    books.forEach((book) => UI.addBookToList(book));
+  }
+
+  // Add book to the List
+  static addBookToList(book) {
+    const list = document.querySelector('#book-list-table');
+    const row = document.createElement('tr');
+    row.innerHTML = `
+          <td id='title-decor'>${book.title}</td>
+          by
+          <td>${book.author}</td>
+          <td>
+            <button><a href='#' id='delete-btn' class='delete'>Delete</a></button>
+          </td>
+        `;
+    list.appendChild(row);
+  }
+
+  // Delete a book from the storage
+  static deleteBook(el) {
+    if (el.classList.contains('delete')) {
+      el.parentElement.parentElement.parentElement.remove();
+    }
+  }
+}
 
 // Event: Display Books
 document.addEventListener('DOMContentLoaded', UI.displayBooks);
