+++ modules/Book.js
@@ -5,4 +5,5 @@ class Book {
   }
 }
 
+
 export default Book;
